server: backup script improved

+++ src/server/backup.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-mv -f -- assigned.000.map assigned.001.map
-mv -f -- complete.000.map complete.001.map
+test -e assigned.000.map && mv -f -- assigned.000.map assigned.001.map
+test -e complete.000.map && mv -f -- complete.000.map complete.001.map
 
-cp -- assigned.map assigned.000.map
-cp -- complete.map complete.000.map
+test -e assigned.map && cp -- assigned.map assigned.000.map
+test -e complete.map && cp -- complete.map complete.000.map
